Retire the schedule pool memo and SessionCache
The session pool cache memoized enumerated prescans keyed by every fact
identity rightly excludes — hints, pins, samples, the split receipt, the
spelled key vocabulary — and each key term was an invariant to defend and a
bug surface (a missed term silently serves a wrong pool). Every schedule()
call now enumerates its own prescan; _Pool, _Draw, the consult/put, and
SessionCache itself (the pool memo was its only tenant) are deleted, along
with Context.session_cache.

Fork.pool_id survives as a STAMP, not a cache key — the variant key + hints +
pins + sample identity — keying the greedy decision memo and seeding the
budgeted descent's deterministic draw. Both consumers fail safe on anything
the stamp cannot see: a replayed row that no longer decodes re-decides. The
receipt and the spelled-vocabulary terms leave the stamp for exactly that
reason. test_schedule_pool_cache.py becomes test_schedule_space_separation.py,
keeping the enumeration-semantics pins (pin-fingerprint gate, transposed
extents, split-dim store, the α-renamed twin's own vocabulary) and dropping
the memo-behavior tests with the memo.
