+++ emmy/compiler/ir/base.py
@@ -149,8 +149,8 @@ class Op:
     # ``identity_key(with_io=True, with_knobs=True)`` is the variant key (the search tree,
     # measurement stores — knobs included because same-body / different-knobs variants must not
     # collide with their parent, and a measurement belongs to (kernel, knob row)). There is
-    # deliberately NO schedule-space key here: the pool memo digest is scheduler plumbing,
-    # minted at the one site that knows every enumeration input (``lowering/tile/_schedule``).
+    # deliberately NO schedule-space key here: the enumeration stamp is scheduler plumbing,
+    # minted at its one site (``lowering/tile/_schedule``).
     # A fact a schedule reads that neither the body nor the io carries is a modeling gap to fix
     # there, never a side-channel fingerprint. ---- #
