+++ emmy/compiler/pipeline/passes/ARCHITECTURE.md
@@ -243,27 +243,14 @@ from a kernel-set decision (`search/golden_eval` filters on it). The same reason
 the sites' own atoms, not off the rows, so a pin naming the scalar tier cannot erase "tensor cores were on offer here"
 from the rows it does enumerate.
 
-**The pool memo and the sampled draw.** The prescan — each node's option list — is memoized in the Context's
-session cache, keyed by every enumeration input: the kernel's the deploy identity (`identity_key(with_io=True)`) (the canonical body + io —
-extents, stores, dtypes and shapes all live there), the knobs, the symbolic-dim hints, the live
-env-pin fingerprint, and two facts the walk consumes directly and therefore keys explicitly — the split receipt
-(`carries_partition`, which strips a `REDUCE` pin's `g`-half where a receipt-free twin must raise) and the spelled
-key vocabulary (the decided-empty OFF map the rows decode under) — so that soundness never rides on how the term
-digest happens to serialize the `compare=False` `Axis.window` or the recognition-canonical axis names, which do
-cover both today — and, when sampling, the sample's identity; target facts need no key
-part because the cache lives ON the Context and one instance never spans two fact sets. What is shared is
-immutable and op-independent — frozen options over read-only knob mappings; options are a function of the node
-and the live pins — so a hit replays the walk over the memo, and every leaf row is a fresh dict the CURRENT
-kernel's own materialization decodes. The memo sits below the search policies — greedy and MCTS hit it alike —
-and holds no ranking and consults no evidence; a prescan that raises (a pin naming nothing) is never memoized.
-The composed digest is also the pool's MINTED identity: `schedule()` stamps it on the shared walk state, every
-Fork of the tree exposes it as `pool_id`, and consumers (the greedy decision memo) key on the stamp instead of
-re-deriving a weaker identity — one minting site, so a decision memo and this pool memo cannot key differently.
-Offline sampling (`emmy fit`, via `ctx.pool_sample`) is the one path that does not return the lazy fork: the
-walk's leaf stream is sampled by single-pass reservoir sampling (`search/pool.py` — nothing proportional to the
-pool is ever retained, and the exact pool size is known when the stream ends), and the drawn complete rows ride
-the memo beside that exact total, keyed apart by the sample's identity, so a sampled Context sharing a session
-cache with a live one can never serve it a draw.
+**The prescan and the sampled draw.** The prescan — each node's option list — is computed fresh per
+`schedule()` call: options are a pure function of the node and the live pins. (The session pool memo that
+cached prescans across kernels is retired: its key had accreted every fact identity rightly excludes —
+hints, pins, samples, the split receipt, the spelled vocabulary — and each was an invariant to defend. The
+`Fork.pool_id` stamp survives as the variant key + hints + pins + sample identity, keying the greedy
+decision memo and seeding the budgeted descent, with drift failing safe into a re-decide.) Under
+`ctx.pool_sample` the walk's leaf stream is reservoir-sampled and the exact pool size reported through
+`sample.totals`.
 
 **Cost is per kernel; a kernel SET is a sum.** A schedule fork picks one alternative and its cost is that
 alternative's latency. A cut's — and a cross-CTA split's — cost is the minimum sum over the kernels it produces,
